chubby/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
torchvision==0.14.0+cpu
flask
fastai
Pillow
gunicorn
asyncio
aiohttp